// web/src/shared/api/generated/model/subjectExamBlueprintsResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ExamBlueprintResult } from './examBlueprintResult';
import type { ExamTypeCountResult } from './examTypeCountResult';
import type { UnitExamBlueprintResult } from './unitExamBlueprintResult';

export interface SubjectExamBlueprintsResult {
  subjectId: string;
  subjectName: string;
  defaultBlueprint: null | ExamBlueprintResult;
  servable: ExamTypeCountResult[];
  units: UnitExamBlueprintResult[];
}